FilterHelper.py: Move filter code to FilterHelper
This avoids annotating filter functions with filter.

Closes #4494
