+++ coalib/parsing/FilterHelper.py
@@ -4,7 +4,6 @@ from coalib.parsing.filters.LanguageFilter import language_filter
 from coalib.parsing.filters.CanDetectFilter import can_detect_filter
 from coalib.parsing.filters.CanFixFilter import can_fix_filter
 
-
 available_filters = {'language': language_filter,
                      'can_detect': can_detect_filter,
                      'can_fix': can_fix_filter}
@@ -14,7 +13,20 @@ def is_valid_filter(filter):
     return filter in available_filters
 
 
-def apply_filter(filter_name, filter_args, all_bears=None):
+def apply_filter(filter_name, filter_args, filter_function, all_bears=None):
+    def filter_section_bears(bears, args):
+        return {section:
+                tuple(bear for bear in bears[section]
+                      if filter_function(bear, args))
+                for section in bears}
+
+    def filter_wrapper(section_to_bears_dict, args):
+        args = {arg.lower() for arg in args}
+        local_bears, global_bears = section_to_bears_dict
+        local_bears = filter_section_bears(local_bears, args)
+        global_bears = filter_section_bears(global_bears, args)
+        return local_bears, global_bears
+
     if all_bears is None:
         from coalib.settings.ConfigurationGathering import (
             get_all_bears)
@@ -27,7 +39,8 @@ def apply_filter(filter_name, filter_args, all_bears=None):
                                              available_filters))))
     if not filter_args or len(filter_args) == 0:
         return all_bears
-    return available_filters[filter_name](all_bears, filter_args)
+    return available_filters[filter_name](all_bears, filter_args),
+    filter_wrapper
 
 
 def apply_filters(filters, bears=None):
